requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.43
sqlmodel==0.0.24
starlette==0.47.2
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
